♻️ Fix all build errors, wire orphan modules, clean codebase ♻️
Resolves 20+ compile errors inherited from prior sessions, wires all
orphaned UI modules into the module tree, removes dead stubs, and adds
project configuration files. Build now passes `cargo check` with 0
errors.
- AI.md: update §22 — build is now error-free as of 2026-06-13
- Cargo.toml: lto="fat", strip="symbols"; add [[test]] entries for all
integration and unit test files; keep russh-sftp at 2.0
- deny.toml: new — cargo-deny config (license allowlist, ban openssl,
allow crates.io only)
- rust-toolchain.toml: new — pin stable channel, rustfmt+clippy, x86_64
and aarch64 targets
- src/app.rs: fix TabBar/Toolbar/StatusBar imports now that components.rs
defines them
- src/bin/ssh_test.rs: fix E0308 — PathBuf→&str conversion for key path
- src/config/mod.rs: rename Settings→AppConfig to eliminate naming
collision with storage::settings::Settings; remove #[allow(dead_code)]
- src/config/themes.rs: fix sed-corrupted format string (comma outside
braces)
- src/lib.rs: export AppConfig instead of Settings; re-export AppConfig
and Theme from config module
- src/main.rs: fix platform and module imports
- src/platform/{bsd,linux,macos,windows}.rs: fix dead_code suppressors
- src/sftp/browser.rs: remove unused anyhow::Result import
- src/sftp/mod.rs: remove #![allow(dead_code)], remove pub mod operations
and its re-exports
- src/sftp/operations.rs: deleted — all 8 methods were TODO stubs
- src/sftp/transfer.rs: remove unused anyhow::Result and Path imports
- src/ssh/active_session.rs: fix russh 0.40 Handler::check_server_key
signature (self by value → Result<(Self, bool), _>)
- src/ssh/config_parser.rs: remove unused anyhow macro and PathBuf import
- src/ssh/connection.rs: fix russh 0.40 API drift
- src/ssh/mod.rs: remove #![allow(dead_code)] and unused import attrs
- src/terminal/buffer.rs: fix TerminalSize{cols,rows} u16→usize casts in
new() and resize()
- src/terminal/emulator.rs: rewrite as facade over TerminalParser +
TerminalRenderer; add process(), size(), resize(), render(), buffer(),
scroll_to_bottom(), with_config()
- src/terminal/mod.rs: wire buffer/cell/parser/renderer orphan modules;
add Color, TerminalSize, Terminal type alias, export RendererConfig and
CursorStyle
- src/terminal/parser.rs: fix buffer.size().rows usize→u16 cast in 'r'
CSI handler
- src/terminal/renderer.rs: fix raw_scroll_delta→scroll_delta (egui 0.25)
- src/ui/components.rs: add TabBar, TabBarAction, Toolbar, ToolbarAction,
StatusBar structs required by app.rs; remove #![allow(dead_code)]
- src/ui/keyboard.rs: remove unused Modifiers import
- src/ui/mod.rs: wire main_window, tab, tab_manager orphan modules; add
pub use re-exports; remove TODO comment
- src/ui/screens/connection_editor.rs: rewrite — replace deleted
connection_manager import with inline ConnectionProfile + ProfileAuthType
definitions; fix egui::Margin::symmetric→vec2; remove TODO comments
- src/ui/screens/connection_list.rs: fix sed-corrupted format strings
- src/ui/screens/connection_manager.rs: deleted — duplicate of
connection_list.rs (322 lines)
- src/ui/screens/forwarding_screen.rs: fix sed-corrupted format string
- src/ui/screens/mod.rs: wire connection_editor and terminal_view orphan
modules
- src/ui/screens/settings.rs: deleted — duplicate of settings_screen.rs
(757 lines)
- src/ui/screens/settings_screen.rs: remove unused BellStyle/CursorStyle
imports
- src/ui/screens/sftp_browser.rs: deleted — duplicate of sftp_browser_ui.rs
(669 lines)
- src/ui/screens/sftp_browser_ui.rs: remove SftpOperations import/field;
fix sed-corrupted format strings
- src/utils/errors.rs: fix unused import suppressors
- src/utils/logging.rs: fix sed-corrupted string literals (whitespace
stripped by prior session's sed -i)

AI.md
AUDIT.AI.md
Cargo.toml
deny.toml
rust-toolchain.toml
src/app.rs
src/bin/ssh_test.rs
src/config/mod.rs
src/config/themes.rs
src/lib.rs
src/main.rs
src/platform/bsd.rs
src/platform/linux.rs
src/platform/macos.rs
src/platform/windows.rs
src/sftp/browser.rs
src/sftp/mod.rs
src/sftp/operations.rs
src/sftp/transfer.rs
src/ssh/active_session.rs
src/ssh/config_parser.rs
src/ssh/connection.rs
src/ssh/mod.rs
src/terminal/buffer.rs
src/terminal/emulator.rs
src/terminal/mod.rs
src/terminal/parser.rs
src/terminal/renderer.rs
src/ui/components.rs
src/ui/keyboard.rs
src/ui/mod.rs
src/ui/screens/connection_editor.rs
src/ui/screens/connection_list.rs
src/ui/screens/connection_manager.rs
src/ui/screens/forwarding_screen.rs
src/ui/screens/mod.rs
src/ui/screens/settings.rs
src/ui/screens/settings_screen.rs
src/ui/screens/sftp_browser.rs
src/ui/screens/sftp_browser_ui.rs
src/utils/errors.rs
src/utils/logging.rs

+++ src/terminal/parser.rs
@@ -206,7 +206,7 @@ impl<'a> Perform for TerminalPerformer<'a> {
             }
             'r' => {
                 let top = param(0, 1).saturating_sub(1) as usize;
-                let bottom = param(1, self.buffer.size().rows).saturating_sub(1) as usize;
+                let bottom = param(1, self.buffer.size().rows as u16).saturating_sub(1) as usize;
                 self.buffer.set_scroll_region(top, bottom);
             }
             's' => {
